move match plan compilation up
